# mobile/backend/docs/dating-api-spec.yaml
openapi: "3.0.3"
info:
  title: Dating API
  version: "1.0.0"
  description: |
    REST API для dating-приложения. Монорепо: backend (Go/chi) + frontend (React SPA).

    **Формат всех ответов:**
    ```json
    { "data": <payload | null>, "error": <"сообщение" | null> }
    ```

    **Rate limits:**
    - `/api/auth/*` — 5 rps, burst 10 (per IP)
    - `/api/*` (защищённые) — 100 rps, burst 200 (per IP)

    **Body limit:** 1 МБ на все запросы.
servers:
  - url: http://localhost:8080
    description: Локальная разработка
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access-токен, выданный при регистрации, логине или OAuth.
        TTL по умолчанию: 15 минут.
  schemas:
...
    Envelope:
      type: object
      required: [data, error]
      properties:
        data:
          description: Полезная нагрузка (null при ошибке)
        error:
          type: string
          nullable: true
          description: Сообщение об ошибке (null при успехе)

    # ── Пользователь ──────────────────────────────────────────────────────────
    ProfileResponse:
      type: object
      description: Публичный профиль пользователя, возвращаемый API.
      required: [id, name, age, bio, interests, location, photo, birthDate]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: display_name из БД
          example: "Алексей"
        age:
          type: integer
          description: Вычисляется из birth_date; 0 если дата не задана
          example: 28
        bio:
          type: string
          description: Описание профиля (пустая строка если не задано)
          example: "Люблю горы и кофе"
        interests:
          type: array
          items:
            type: string
          description: Список интересов (JSONB → []string)
          example: ["hiking", "coffee", "dev"]
        location:
          type: string
          description: Город/локация (пустая строка если не задано)
          example: "Москва"
        photo:
          type: string
          description: avatar_url (пустая строка если не задано)
          example: "https://cdn.example.com/media/abc.jpg"
        birthDate:
          type: string
          format: date
          description: Дата рождения в формате YYYY-MM-DD (пустая строка если не задана)
          example: "1996-03-15"
        personality:
          description: Произвольный JSONB-объект данных личности (опционально)
          nullable: true
          example: { "type": "INTJ", "traits": ["analytic"] }
        astrology:
          description: Произвольный JSONB-объект астрологических данных (опционально)
          nullable: true
          example: { "sign": "Рыбы", "element": "Вода" }
        numerology:
          description: Произвольный JSONB-объект нумерологических данных (опционально)
          nullable: true
          example: { "lifePathNumber": 7 }
...
    ErrorResponse:
      type: object
      description: Оболочка при ошибке.
      required: [data, error]
      properties:
        data:
          nullable: true
          example: null
        error:
          type: string
          example: "неверный email или пароль"
paths:
...
  /api/users/me:
    get:
      tags: [Users]
      summary: Получить профиль текущего пользователя
      description: Возвращает ProfileResponse для авторизованного пользователя.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Профиль текущего пользователя
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден в БД
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags: [Users]
      summary: Обновить профиль текущего пользователя
      description: |
        Частичное обновление профиля (все поля опциональны).
        Передавайте только те поля, которые нужно обновить.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: |
                Все поля опциональны. Поля с omitempty не передавайте,
                чтобы не перезаписать их пустыми значениями.
              properties:
                name:
                  type: string
                  nullable: true
                  description: Новое отображаемое имя
                  example: "Алекс"
                bio:
                  type: string
                  nullable: true
                  description: Новое описание профиля
                  example: "Люблю горы"
                location:
                  type: string
                  nullable: true
                  description: Новая локация
                  example: "Санкт-Петербург"
                avatarUrl:
                  type: string
                  nullable: true
                  description: URL аватара (обычно fileUrl из /api/upload)
                  example: "https://cdn.example.com/media/avatar.jpg"
                interests:
                  type: array
                  items:
                    type: string
                  description: Полный новый список интересов (заменяет старый)
                  example: ["hiking", "coffee"]
                personality:
                  description: Произвольный JSONB-объект данных личности
                  example: { "type": "INTJ" }
                astrology:
                  description: Произвольный JSONB-объект астрологических данных
                  example: { "sign": "Рыбы" }
                numerology:
                  description: Произвольный JSONB-объект нумерологических данных
                  example: { "lifePathNumber": 7 }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProfileResponse"
        "400":
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Получить профиль пользователя по ID
      description: Возвращает публичный ProfileResponse для любого пользователя по UUID.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID пользователя
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Профиль пользователя
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProfileResponse"
        "400":
          description: Невалидный UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ════════════════════════════════════════════════════════════════════════════
  # CONVERSATIONS
  # ════════════════════════════════════════════════════════════════════════════
...